@@ -4,7 +4,7 @@ buildscript {
         pluginVersion = '4.0.0-SNAPSHOT'
         
         // Java compile
-        javaTargetVersion = JavaVersion.VERSION_1_7
+        javaTargetVersion = JavaVersion.VERSION_1_8
     }
     repositories {
         jcenter()
